rename analysis to models and reorganize
